Refactor README generator for improved organization and error handling
